+++ pyodsp/alg/bm/logger.py
@@ -1,5 +1,4 @@
 import logging
-from typing import List
 
 
 class BmLogger:
