+++ pyodsp/dec/dd/alg_root_bm.py
@@ -11,7 +11,7 @@ from .message import DdDnMessage, DdFinalDnMessage, DdInitDnMessage, DdFinalUpMe
 from .master_creator import MasterCreator
 from .mip_heuristic_root import IMipHeuristicRoot
 from pyodsp.alg.bm.bm import BundleMethod
-from pyodsp.alg.pbm.pbm import ProximalBundleMethod
+from pyodsp.alg.bm.pbm import ProximalBundleMethod
 from pyodsp.alg.cuts import CutList
 from pyodsp.alg.cuts_manager import CutInfo
 from pyodsp.alg.params import BM_DUMMY_BOUND
